fix build error by moving out shared page component from template to standard component.
See: https://github.com/gatsbyjs/gatsby/issues/1846
